Fix rollback issue when the user plugin gets refreshed first. (#36)

+++ updates/user_add_location_fields.php
@@ -20,7 +20,7 @@ class UserAddLocationFields extends Migration
 
     public function down()
     {
-        if (Schema::hasTable('users')) {
+        if (Schema::hasTable('users') && Schema::hasColumns('users', ['state_id', 'country_id'])) {
             Schema::table('users', function ($table) {
                 $table->dropColumn(['state_id', 'country_id']);
             });
